Update index.html
The first thing I did was go through my code and change all the functions, variables and questions to resemble the two new diseases I am comparing: mono and strep. I replaced all the "cold and allergy" functions to say "mono and strep", and changed the questions to diagnose the likelihood of having those diseases.